+++ javasdk/src/main/resources/plugin_settings_file.json
