+++ javasdk/src/main/resources/vapil_settings_file.json
