+++ pom.xml
@@ -9,7 +9,7 @@
     <version>1.0-SNAPSHOT</version>
 
       <properties>
-          <vault.sdk.version>[25.3.0-release0, 25.3.0-release1000000]</vault.sdk.version>
+          <vault.sdk.version>[26.1.2-release0, 26.1.2-release1000000]</vault.sdk.version>
       </properties>
 
     <repositories>
@@ -65,8 +65,8 @@
                 <artifactId>vaultjavasdk-maven-plugin</artifactId>
                 <version>24.1.0</version>
                 <configuration>
-                    <pluginSettingsFilePath>${project.basedir}${file.separator}src${file.separator}main${file.separator}resources${file.separator}plugin_settings_file.json</pluginSettingsFilePath>
-                    <vapilSettingsFilePath>${project.basedir}${file.separator}src${file.separator}main${file.separator}resources${file.separator}vapil_settings_file.json</vapilSettingsFilePath>
+                    <pluginSettingsFilePath>${project.basedir}${file.separator}javasdk${file.separator}src${file.separator}main${file.separator}resources${file.separator}plugin_settings_file.json</pluginSettingsFilePath>
+                    <vapilSettingsFilePath>${project.basedir}${file.separator}javasdk${file.separator}src${file.separator}main${file.separator}resources${file.separator}vapil_settings_file.json</vapilSettingsFilePath>
                 </configuration>
             </plugin>
         </plugins>
